requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
pandas
python-dotenv
streamlit #optional